Remove BingoCard component
